requirements.txt
dnspython==2.8.0
pyshark=
psutil=
asyncio=